django-server/decendraline/requirements.txt
aiohttp @ file:///D:/bld/aiohttp_1696951890432/work
aiosignal @ file:///home/conda/feedstock_root/build_artifacts/aiosignal_1667935791922/work
airpyllution==2.0.1
altair==4.2.2
altair-data-server==0.4.1
altair-viewer==0.4.0
annotated-types==0.6.0
anyio==3.7.1
asgiref @ file:///home/conda/feedstock_root/build_artifacts/asgiref_1685243452975/work
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1683424013410/work
black==21.12b0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.1
DateTime==4.9
Django @ file:///home/conda/feedstock_root/build_artifacts/django_1696523235836/work
django-cors-headers @ file:///home/conda/feedstock_root/build_artifacts/django-cors-headers_1689682275336/work
djangorestframework @ file:///home/conda/feedstock_root/build_artifacts/djangorestframework_1663854154474/work
entrypoints==0.4
frozenlist @ file:///D:/bld/frozenlist_1695377816968/work
greenlet==3.0.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
langchain==0.0.312
langsmith==0.0.43
MarkupSafe==2.1.3
marshmallow==3.20.1
mock==4.0.3
multidict @ file:///D:/bld/multidict_1696716185902/work
mypy-extensions==1.0.0
numpy==1.26.0
openai==0.28.1
packaging==23.2
pandas==1.5.3
pathspec==0.11.2
platformdirs==3.11.0
plotly==5.17.0
portpicker==1.6.0
psutil==5.9.5
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1693930252784/work
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
responses==0.17.0
rpds-py==0.10.6
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
sqlparse @ file:///home/conda/feedstock_root/build_artifacts/sqlparse_1681817562700/work
tenacity==8.2.3
tomli==1.2.3
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1695040754690/work
urllib3==2.0.6
vega-datasets==0.9.0
wheel==0.41.2
yarl @ file:///D:/bld/yarl_1696732602322/work
zope.interface==6.1
